requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
python-dotenv==1.0.0
streamlit==1.29.0
openai>=1.10.0
dashscope==1.14.1


